Add xField in a separate files
