+++ libNeonDomain/include/Neon/domain/internal/mGrid/mField.h
@@ -6,6 +6,8 @@
 #include "Neon/domain/internal/bGrid/bField.h"
 #include "Neon/domain/internal/mGrid/mPartition.h"
 
+#include "Neon/domain/internal/mGrid/xField.h"
+
 namespace Neon {
 
 enum struct MultiResCompute
@@ -24,160 +26,6 @@ class bGrid;
 namespace Neon::domain::internal::mGrid {
 class mGrid;
 
-
-template <typename T, int C = 0>
-class xField : public Neon::domain::interface::FieldBaseTemplate<T,
-                                                                 C,
-                                                                 Neon::domain::internal::bGrid::bGrid,
-                                                                 mPartition<T, C>,
-                                                                 int>
-
-{
-   public:
-    using Field = typename Neon::domain::internal::bGrid::bField<T, C>;
-    using Partition = typename Neon::domain::internal::mGrid::mPartition<T, C>;
-    using Grid = Neon::domain::internal::bGrid::bGrid;
-
-
-    xField() = default;
-
-    xField(const std::string&         name,
-           const Grid&                grid,
-           int                        cardinality,
-           T                          outsideVal,
-           Neon::DataUse              dataUse,
-           const Neon::MemoryOptions& memoryOptions)
-        : Neon::domain::interface::FieldBaseTemplate<T, C, Grid, Partition, int>(&grid,
-                                                                                 name,
-                                                                                 "xbField",
-                                                                                 cardinality,
-                                                                                 outsideVal,
-                                                                                 dataUse,
-                                                                                 memoryOptions,
-                                                                                 Neon::domain::haloStatus_et::ON)
-    {
-        mData = std::make_shared<Data>();
-        mData->field = grid.newField(name, cardinality, outsideVal, dataUse, memoryOptions);
-    }
-
-
-    auto isInsideDomain(const Neon::index_3d& idx) const -> bool final
-    {
-        return this->mData->field.isInsideDomain(idx);
-    }
-
-    auto getReference(const Neon::index_3d& idx, const int& cardinality) -> T& final
-    {
-        return this->operator()(idx, cardinality);
-    }
-
-
-    auto haloUpdate(Neon::set::HuOptions& opt) const -> void final
-    {
-        mData->field.haloUpdate(opt);
-    }
-
-    auto haloUpdate(Neon::set::HuOptions& opt) -> void final
-    {
-        mData->field.haloUpdate(opt);
-    }
-
-    auto operator()(const Neon::index_3d& idx, const int& cardinality) const -> T final
-    {
-        return mData->field.getReference(idx, cardinality);
-    }
-
-    auto operator()(const Neon::index_3d& idx,
-                    const int&            cardinality) -> T&
-    {
-        return mData->field.getReference(idx, cardinality);
-    }
-
-
-    auto getPartition(const Neon::DeviceType& devType,
-                      const Neon::SetIdx&     idx,
-                      const Neon::DataView&   dataView) const -> const Partition&
-    {
-        if (devType == Neon::DeviceType::CUDA) {
-            return mData->mPartitions[PartitionBackend::gpu][Neon::DataViewUtil::toInt(dataView)][idx];
-        } else {
-            return mData->mPartitions[PartitionBackend::cpu][Neon::DataViewUtil::toInt(dataView)][idx];
-        }
-    }
-
-    auto getPartition(const Neon::DeviceType& devType,
-                      const Neon::SetIdx&     idx,
-                      const Neon::DataView&   dataView) -> Partition&
-    {
-        if (devType == Neon::DeviceType::CUDA) {
-            return mData->mPartitions[PartitionBackend::gpu][Neon::DataViewUtil::toInt(dataView)][idx];
-        } else {
-            return mData->mPartitions[PartitionBackend::cpu][Neon::DataViewUtil::toInt(dataView)][idx];
-        }
-    }
-
-    auto getPartition(Neon::Execution       exec,
-                      Neon::SetIdx          idx,
-                      const Neon::DataView& dataView) const -> const Partition& final
-    {
-
-        if (exec == Neon::Execution::device) {
-            return getPartition(Neon::DeviceType::CUDA, idx, dataView);
-        }
-        if (exec == Neon::Execution::host) {
-            return getPartition(Neon::DeviceType::CPU, idx, dataView);
-        }
-
-        NEON_THROW_UNSUPPORTED_OPERATION("xField::getPartition() unsupported Execution");
-    }
-
-
-    auto getPartition(Neon::Execution       exec,
-                      Neon::SetIdx          idx,
-                      const Neon::DataView& dataView) -> Partition& final
-    {
-        if (exec == Neon::Execution::device) {
-            return getPartition(Neon::DeviceType::CUDA, idx, dataView);
-        }
-        if (exec == Neon::Execution::host) {
-            return getPartition(Neon::DeviceType::CPU, idx, dataView);
-        }
-
-        NEON_THROW_UNSUPPORTED_OPERATION("xField::getPartition() unsupported Execution");
-    }
-
-    auto updateIO(int streamId) -> void
-    {
-        mData->field.updateIO(streamId);
-    }
-
-    auto updateCompute(int streamId) -> void
-    {
-        mData->field.updateCompute(streamId);
-    }
-
-    virtual ~xField() = default;
-
-
-    enum PartitionBackend
-    {
-        cpu = 0,
-        gpu = 1,
-    };
-    struct Data
-    {
-        Field field;
-
-        std::array<
-            std::array<
-                Neon::set::DataSet<Partition>,
-                Neon::DataViewUtil::nConfig>,
-            2>  //2 for host and device
-            mPartitions;
-    };
-
-    std::shared_ptr<Data> mData;
-};
 
 template <typename T, int C = 0>
 class mField
